extract al_vm: the language-agnostic runtime substrate
New crate al_vm holds everything the VM needs at runtime and nothing the
language front end owns: the bytecode ISA (Op/Instruction/Function/Program),
the NaN-boxed value representation and heap object layouts, the per-process
heap, the frozen area, the persistent seq/hamt collections, the native-entry
table, and the cranelift RC emission (native_rc). TypeId, FuncIdx, and
VariantTemplate move too - they are runtime identities the VM stores.

al_core depends on al_vm and re-exports the moved items at their historical
paths (al_core::bytecode::value, al_core::heap, ...), so the compiler side
is unchanged. builtin_op (the @vm intrinsic name table) stays in al_core:
the strings are AL stdlib conventions, not VM facts.

Cargo now enforces the boundary: al_vm cannot name an AST node or a type
scheme, because depending on al_core would be a cycle.
