+++ crates/al_core/src/lib.rs
@@ -9,11 +9,10 @@
         clippy::unimplemented,
     )
 )]
-// Unsafe code is confined to designated modules (bytecode::value, heap::proc_heap,
-// frozen, plus the scoped allows on bytecode::fetch (in bytecode::mod),
-// bytecode::scratch, and bytecode::native's entry-pointer transmute) — each
-// carries its own `allow(unsafe_code)` and justification. Everything else is
-// compiler-enforced safe.
+// All unsafe code lives in `al_vm` (the runtime crate); this crate — the
+// language front end — is compiler-enforced safe, with the one scoped allow
+// on `core_ir::clif`'s tests, which drive JIT'd code through a C-ABI mock
+// runtime.
 #![deny(unsafe_code)]
 
 pub mod ast;
@@ -21,8 +20,6 @@ pub mod bytecode;
 pub mod core_ir;
 pub mod diagnostic;
 pub mod formatter;
-pub mod frozen;
-pub mod heap;
 pub mod module;
 pub mod parser;
 pub mod precompile;
@@ -31,18 +28,17 @@ pub mod scanner;
 pub mod span;
 pub mod static_ir;
 pub mod term;
-pub mod tivec;
 pub mod token;
 pub mod type_def;
 pub mod typed_ir;
 pub mod types;
 
+// The runtime substrate — values, heap, frozen area, index vectors — lives in
+// `al_vm`; re-exported at the historical paths so `al_core::heap` etc. keep
+// naming the one shared definition.
+pub use al_vm::{assert_send, assert_send_sync, frozen, heap, tivec};
+
 pub use bytecode::{CtorRef, PreludeBindings, TypeRef};
 pub use precompile::{PrecompileOutput, precompile_stdlib};
 pub use static_ir::{StaticStdlib, VariantTemplate};
 pub use type_def::TypeId;
-
-/// Compile-time assertion that `T: Send`. Use as `const _: () = assert_send::<T>();`.
-pub const fn assert_send<T: Send>() {}
-/// Compile-time assertion that `T: Send + Sync`.
-pub const fn assert_send_sync<T: Send + Sync>() {}
