+++ crates/al_vm/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "al_vm"
+version = "0.1.0"
+edition = "2024"
+
+[features]
+# Instruments `ProcHeap::alloc_object` with a per-thread counter so tests can
+# assert net allocations (Perceus reuse). Enabled only from `al`'s
+# [dev-dependencies]; with `resolver = "2"` that keeps it out of `cargo build`,
+# so the release binary carries no counter.
+#
+# It has to be a feature rather than `cfg(test)`: the tests that read it are
+# `al`'s integration tests, where `al_vm` is a plain dependency and its
+# `cfg(test)` is false.
+alloc-counter = []
+
+[dependencies]
+indexmap = { workspace = true }
+cranelift-codegen = { workspace = true }
+cranelift-frontend = { workspace = true }
+cranelift-module = { workspace = true }
+cranelift-jit = { workspace = true }
+cranelift-native = { workspace = true }
+smallvec = "1.15"
+libmimalloc-sys = { version = "0.1.49", features = ["v2"] }
