+++ crates/al_vm/src/bytecode/hamt.rs
