+++ crates/al_vm/src/bytecode/scratch.rs
