+++ crates/al_vm/src/heap/mod.rs
